refactor: 리팩토링, 미사용 코드 삭제, 폴더명 변경 (#181)
* refactor: LayoutContextType 분리해 가독성 개선

* rername: services -> apis로 이름 변경

* remove: 미사용 코드 삭제

* rename: apis -> api, libs -> lib
