+++ src/api/score.ts
