+++ src/app/community/[boardCode]/[postId]/modify/PostModifyForm.tsx
@@ -3,8 +3,7 @@
 import { useRouter } from "next/navigation";
 import { useEffect, useRef, useState } from "react";
 
-import { updatePostApi } from "@/services/community";
-
+import { updatePostApi } from "@/api/community";
 import { IconArrowBackFilled, IconImage, IconPostCheckboxFilled, IconPostCheckboxOutlined } from "@/public/svgs";
 
 type PostModifyFormProps = {
