+++ src/app/login/kakao/callback/KakaoLoginCallbackPage.tsx
@@ -5,12 +5,12 @@ import { useEffect } from "react";
 
 import axios from "axios";
 
-import { kakaoAuthApi } from "@/services/auth";
 import { saveAccessToken, saveRefreshToken } from "@/utils/localStorage";
 
 import CloudSpinnerPage from "@/components/loading/CloudSpinnerPage";
 import SignupSurvey from "@/components/login/signup/SignupSurvey";
 
+import { kakaoAuthApi } from "@/api/auth";
 import { useLayout } from "@/context/LayoutContext";
 
 const KakaoLoginCallbackPage = () => {
