+++ src/app/university/[id]/page.tsx
@@ -2,8 +2,6 @@ import { Metadata, ResolvingMetadata } from "next";
 import Head from "next/head";
 import { notFound } from "next/navigation";
 
-import { getUniversityDetailPublicApi } from "@/services/university";
-
 import TopDetailNavigation from "@/components/layout/TopDetailNavigation";
 
 import CollegeBottomSheet from "./CollegeBottomSheet";
@@ -11,6 +9,8 @@ import CollegeDetail from "./CollegeDetail";
 
 import { Review } from "@/types/review";
 import { University } from "@/types/university";
+
+import { getUniversityDetailPublicApi } from "@/api/university";
 
 export const revalidate = 60; // ISR 재생성 주기 설정
