+++ src/components/home/UniversityList.tsx
@@ -4,7 +4,6 @@ import Image from "next/image";
 import Link from "next/link";
 import { useEffect, useState } from "react";
 
-import { getUniversityListPublicApi } from "@/services/university";
 import { shortenLanguageTestName } from "@/utils/universityUtils";
 
 import ButtonTab from "@/components/ui/ButtonTab";
@@ -12,6 +11,8 @@ import CheveronRightFilled from "@/components/ui/icon/ChevronRightFilled";
 
 import { languageTestShortMapping } from "@/types/score";
 import { ListUniversity, regionMapping } from "@/types/university";
+
+import { getUniversityListPublicApi } from "@/api/university";
 
 const UniversityList = () => {
   const [region, setRegion] = useState<"전체" | "영미권" | "유럽권" | "아시아권">("전체");
