+++ src/pages/search.tsx
@@ -1,11 +1,11 @@
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 
-import { getPopularKeywordsPublicApi, postSearchKeywordPublicApi } from "@/services/keyword";
-
 import TopDetailNavigation from "@/components/layout/TopDetailNavigation";
 import CollegeSearchBar from "@/components/search/college-search-bar";
 import CollegeSearchField from "@/components/search/college-search-field";
+
+import { getPopularKeywordsPublicApi, postSearchKeywordPublicApi } from "@/api/keyword";
 
 const HomeSearchPage = () => {
   const router = useRouter();
